add opabigint
opabigint is now the general interface to access the underlying
biginteger library... will support libtommath, mbedtls, or gmp now. More
libraries can be added in the future. This should unify and simplify
some of the code. Also, libtommath's mp_init function was problematic in
that it returned an error code. opabigintInit() does not return an
error! Note that some functionality may not be 100% consistent between
all libraries. The opabigint API needs to be solidified on boundary
cases and test cases created to test all current and new libs. For
example, what happens when opabigintGetMagU64() is called on a bigint
that has more than 64 bits? This should be defined in the API and each
library shim should handle all boundary cases the same.