opacore: rework logging for windows
The main goal of this change is to support logging utf-8 strings to
console on windows. When stdout and stderr are redirected to a file on
windows then things work as expected. However, when stdout/stderr are
printing to the console, utf-8 strings will not be displayed properly.
Rather than using printf/fprintf/etc, this code must convert from utf-8
char* to wchar_t* and then call WriteConsoleW().

Also, when getting an error message from a win32 error code such as
GetLastError(), use FormatMessageW() to print the unicode message in the
user's language.

Code should call opa_printf() or opa_fprintf() rather than
printf/fprintf to make sure utf-8 strings are printed properly on
windows consoles.