opacore: fix console logging on win2k
vsnprintf(NULL, 0, ...) is supposed to return the length required for
printing. However, on win2k it returns -1. Therefore, ws_vfprintf()
fails to print anything on win2k.

This new code works around the issue while also trying to run fast on
newer windows that might return the correct value when the buffer is too
small for vsnprintf().